fix: text outside page when preformatted
When Text is preformatted, it is not wrapped and can go outside the page.
added max-width to preformatted text to prevent this.

Also, removed white-space: nowrap; it was added because there was some
issue with wkhtmltopdf which breaks text.
don't remember when this happens so will revert this if it breaks again.