fix: don't update schema_version on load
schema_version is suppose to indicate new data version.
instead of updating it from db on load we will save it as old_schema_version
As when we save, we are suppose to fix the state that we couldn't
do using patches.
we want to fix the schema and update the schema version on save.