fix: print format set standard to Yes
when Standard is set to Yes, JSON file for print format is created.
however, logic for print format in framework assumed that when standard is set to Yes
HTML file should be loaded from disk. which is incorrect as HTML file is not required
when standard is set to Yes for print designer.

fix for this is raised in Frappe framework.
ref: https://github.com/frappe/frappe/pull/24694

This PR contains client script for Print Designer and patch for print designer formats.