fix: pos print and added trigger_print
added client script to update frappe.util.print function that is used in pos.
Also, added option to trigger print on print view by passing it as URL Parameter.
If trigger_print=1 is passed it will close the window in 5 second after print dialog is closed.
This is added to maintain current behavior.
If this causes issue we can introduce 1 more arg for closing and pass when using frappe.util.print.