feat: add isDynamicHeight property for user control
This property is how user can configure if some element should have fixed height or should be dynamically adjusted.
This will not ensure that it will have dynamically calculated height as it will depend on several factors
This is just to give them control to choose between static and dynamic height.