fix: printview and table header overlap issue
wkhtmltopdf have issue with table header overlapping if it is inside a div with display: inline.

wkhtmltopdf doesn't support flex but it has some old experimental support for https://developer.mozilla.org/en-US/docs/Web/CSS/box-align

As it doesn't support proper implementation of flex it is not used anywhere in print designer.

This PR uses display: -webkit-box; to fix the particular issue.

hopefully can be used to add more features in the future.