feat: add new height option auto (min-height)
if rectangle is only ( basically no element is overlapping with it) or last element inside row or column, row and column layout is computed for children.
- if rectangle is overlapping with other element and we make it relative placement of other element will be broken so for now it is set to compute layout as per above condition

created patch that will convert isDynamicHeight Boolean to heightType with value "fixed" or "auto".

updated default objects and  control in Properties Panel.

changed schema version number to 1.2.0 as schema is changed and is incompatible with old formats (1.1.0)  we need to handle old formats.

updated jinja render template
- modified relative_containers to be recursive
- updated all templates to use heightType properties
- updated logic to include min-height
- added logic for backwards compatibility.
- also introduced dynamic height for static text if it has parseJinja enabled.
- updated rectangle template to be dynamic height

misc:
    is_older_schema now handles really old formats which don't have schema_version
    reverted styles for relative column