fix: don't close the browser if it is in use by another thread
chrome browser is shared per worker, so we should not close it if any of the threads is still using it
add browserId on pdf generation and remove at the end.
close browser only if no browserId is present hence no thread is using it

not saving browser instance (may contain sensitive information) as FrappePDFGenerator instance is shared across threads in worker.