feat: load local resources directly from the file system
we are making request from server to nginx to get the local resources
this is not needed, we can directly load the resources from the file system
this will reduce no of requests on the nginx.

we intercept the request and check if the request path is "assets/" or "files/" then we load the resources from the file system.